+++ CHANGELOG.md
@@ -10,6 +10,7 @@ nav_order: 1
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 
 - Add additional disk space support
+- Update disk space on refresh
 
 ## [3.5.1] - 2022-08-16